refactor : match dir name
