+++ src/hooks/index.ts
@@ -3,4 +3,5 @@ export * from './useDetectResolution';
 export * from './useHandleOpenModal';
 export * from './useViewportWidth';
 export * from './useArticleElement';
-export * from './article/useHandleArticle';
+export * from './article/useResolveSaveHandler';
+export * from './article/useHandleDelete';
